chore: permissions refactor
