+++ src/utils/hooks/usePermissions.ts
@@ -1,0 +1,33 @@
+import { PermissionsAndroid, Platform } from 'react-native';
+
+export const usePermissions = () => {
+  const requestAudioPermissions = async (): Promise<boolean> => {
+    if (Platform.OS === 'android') {
+      try {
+        const grants = await PermissionsAndroid.requestMultiple([
+          PermissionsAndroid.PERMISSIONS.WRITE_EXTERNAL_STORAGE,
+          PermissionsAndroid.PERMISSIONS.READ_EXTERNAL_STORAGE,
+          PermissionsAndroid.PERMISSIONS.RECORD_AUDIO,
+        ]);
+
+        console.log('Permissions status:', grants);
+
+        return (
+          grants['android.permission.WRITE_EXTERNAL_STORAGE'] ===
+            PermissionsAndroid.RESULTS.GRANTED &&
+          grants['android.permission.READ_EXTERNAL_STORAGE'] ===
+            PermissionsAndroid.RESULTS.GRANTED &&
+          grants['android.permission.RECORD_AUDIO'] ===
+            PermissionsAndroid.RESULTS.GRANTED
+        );
+      } catch (err) {
+        console.warn('Permission request failed:', err);
+        return false;
+      }
+    }
+
+    return true;
+  };
+
+  return { requestAudioPermissions };
+};
